Deprecate non-used perf
